+++ 03.asynchronous/src/promise.js
@@ -1,11 +1,15 @@
 #!/usr/bin/env node
-import { runPromise, allPromise, closePromise } from "./db/promiseFunctions.js";
+import {
+  runPromise,
+  allPromise,
+  closePromise,
+} from "../db/promiseFunctions.js";
 import {
   createTableSQL,
-  invalidCreateBooksSQL,
-  invalidGetBooksSQL,
+  createBooksSQL,
+  getBooksSQL,
   deleteTableSQL,
-} from "./db/queries.js";
+} from "../db/queries.js";
 
 const titles = ["書籍1", "書籍2", "書籍3"];
 
@@ -18,18 +22,14 @@ const createBooks = (titles, index = 0) => {
     return Promise.resolve();
   }
 
-  return runPromise(invalidCreateBooksSQL, [titles[index]])
-    .then((result) => {
-      console.log(`ID: ${result.lastID} created.`);
-      return createBooks(titles, index + 1);
-    })
-    .catch((error) => {
-      return Promise.reject(error);
-    });
+  return runPromise(createBooksSQL, [titles[index]]).then((result) => {
+    console.log(`ID: ${result.lastID} created.`);
+    return createBooks(titles, index + 1);
+  });
 };
 
 const getBooks = () => {
-  return allPromise(invalidGetBooksSQL);
+  return allPromise(getBooksSQL);
 };
 
 const displayBooks = (books) => {
@@ -47,12 +47,6 @@ const closeDB = () => {
   return closePromise();
 };
 
-const displayError = (error) => {
-  if (error) {
-    console.error("Error: " + error.message);
-  }
-};
-
 function main() {
   createTable()
     .then(() => {
@@ -66,9 +60,6 @@ function main() {
     })
     .then(() => {
       return deleteTable();
-    })
-    .catch((error) => {
-      displayError(error);
     })
     .finally(() => {
       closeDB();
